requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
psutil==5.9.6
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
paramiko==3.3.1
dnspython==2.4.2
requests==2.31.0
